+++ src/WebJobs.Script.WebHost/Management/LinuxSpecialization/IManagedIdentityTokenProvider.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.WebJobs.Script.WebHost.Management.LinuxSpecialization
+{
+    public interface IManagedIdentityTokenProvider
+    {
+        Task<string> GetManagedIdentityToken(string resourceUrl);
+    }
+}