+++ test/WebJobs.Script.Tests.Integration/Management/KubernetesPodControllerTests.cs
@@ -49,7 +49,10 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
                 StatusCode = HttpStatusCode.OK
             });
 
-            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object), scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(), new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object);
+            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object),
+                scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(),
+                new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object,
+                new Mock<IPackageDownloadHandler>(MockBehavior.Strict).Object);
             var startupContextProvider = new StartupContextProvider(environment, loggerFactory.CreateLogger<StartupContextProvider>());
 
             InstanceManager.Reset();
@@ -104,7 +107,10 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
                 StatusCode = HttpStatusCode.OK
             });
 
-            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object), scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(), new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object);
+            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object),
+                scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(),
+                new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object,
+                new Mock<IPackageDownloadHandler>(MockBehavior.Strict).Object);
             var startupContextProvider = new StartupContextProvider(environment, loggerFactory.CreateLogger<StartupContextProvider>());
 
             InstanceManager.Reset();